fix: remove phpstan baseline, fix all type errors
- Delete phpstan-baseline.neon (no longer suppressing errors)
- Add @return generic PHPDoc to all Eloquent relation methods
  (HasOne, BelongsTo, HasOneThrough) so Larastan can infer proper
  property types instead of returning Model
- Fix $guarded = false → $guarded = [] in Contact, ContactLabel,
  Label, Balance, WalletJournal, WalletTransaction
- Fix $guarded PHPDoc array<string>|bool → array<string> in System, Type
- Fix $dispatchesEvents PHPDoc @var array → @var array<string, class-string>
  in Constellation, Station, Structure, System
- Add @property annotations and make accessor methods protected for
  KillmailItem::has_content and ContactLabel::label_name (Larastan
  modelAppends rule requires protected accessor methods)
- Add Label import and typed local var in ContactLabel::labelName()
- Cast diffInMinutes() to int in UpdateCharacter::calculateInterval()
- Remove nullsafe ?-> on character->name after relation generics applied
- Fix ResolveLocationService: $this->resolvers ??= → = $resolvers ?:
- Replace map+flatten with flatMap in GetMissingLocationFromContracts
- Update rector.php: UP_TO_LARAVEL_110 → UP_TO_LARAVEL_130

+++ src/Jobs/Seatplus/UpdateCharacter.php
@@ -110,6 +110,6 @@ class UpdateCharacter implements ShouldQueue
     {
         $cron = new CronExpression($expression);
 
-        return carbon($cron->getPreviousRunDate())->diffInMinutes($cron->getNextRunDate(null));
+        return (int) carbon($cron->getPreviousRunDate())->diffInMinutes($cron->getNextRunDate(null));
     }
 }
